api/requirements.txt


anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.5
pipreqs==0.4.11
pycodestyle==2.10.0
pycparser==2.21
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette
tomli==2.0.1
typing-extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn>=0.15.0,<0.16.0
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
zipp==3.11.0
fastapi
fastapi_jwt_auth==0.5.0
Pillow==9.4.0
pydantic==1.10.4
pymongo==4.3.3
python-decouple==3.7
python-dotenv==0.21.1
requests==2.28.2
